♻️ Refactor: 기존 play 함수에서 작은 단위로 분리하기
